Simplify and clean up reverseBits.
This PR aims to clean up the reverseBits function.

1. The implementation is more "modern", functional JS, in that it is less algorithmically intensive, simplifying the implementation.
2. The implementation, in my opinion, is cleaner and more digestible.

The disadvantage of this implementation is that it is about 3 times slower than the legacy implementation. This is naturally to be expected due to the expensive `join`, however, noting that the previous implementation was also suboptimal, I think this should be good enough!

The stanford bit-fiddling website has some neat tricks on this.