Simplify 13. Roman to Integer
This PR aims to make the implementation simpler and more modern.

The performance is about the same -- ~220ms for the older implementation and ~240ms for the suggested change. Likely the slowdown is from the lambda stack overhead. The slowdown is not too bad since I've managed to eliminate the redundant if-else.

I suspect for practical complexity to be a bit worse than the previous implementation.

The tradeoff is that the code is safer and way more readable.